Moving iree_aligned_alloc to iree/base/allocator.h. (#6321)
